requirements.txt
mne
tqdm
h5py
xlrd
numpy
scipy
pandas
dataframe_image
scikit-learn
xgboost
seaborn
jupyter
openpyxl
kendall-w
quantities
matplotlib
joypy
ruamel.yaml
biosppy
hrv-analysis
distinctipy
umap-learn
imbalanced-learn
tensorboard
torchinfo
wandb
einops
torchmetrics
coral_pytorch
flirt